programs/reasoning-engine/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "reasoning-engine"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "oracle-shared",
    "redis[hiredis]>=5.0",
    "python-dotenv>=1.0",
    "anthropic>=0.30",
    "openai>=1.0",
    "chromadb>=0.4",
    "apscheduler>=3.10",
]

[tool.hatch.build.targets.wheel]
packages = ["reasoning_engine"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
